--- src/components/common/Input.tsx
@@ -1,20 +1,0 @@
-import { forwardRef } from "react";
-
-export interface InputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {}
-
-const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ type = "text", disabled = false, required = false, ...props }, ref) => {
-    return (
-      <input
-        ref={ref}
-        type={type}
-        disabled={disabled}
-        required={required}
-        {...props}
-      />
-    );
-  }
-);
-
-export default Input;
