--- src/stories/Input.stories.ts
@@ -1,20 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import { Input } from "../components/common";
-
-const meta = {
-  title: "페이먼츠/Input",
-  component: Input,
-  parameters: {
-    layout: "centered",
-  },
-  tags: ["autodocs"],
-  argTypes: {
-    disabled: { control: "boolean" },
-    required: { control: "boolean" },
-  },
-} satisfies Meta<typeof Input>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Primary: Story = {};
