+++ src/styles/input.css
@@ -23,30 +23,6 @@
   color: #525252;
 }
 
-.input-basic {
-  background-color: #ecebf1;
-  height: 45px;
-  width: 100%;
-  text-align: center;
-  outline: 2px solid transparent;
-  outline-offset: 2px;
-  border-color: #9ca3af;
-  border: none;
-  border-radius: 0.25rem;
-}
-
-.input-underline {
-  text-align: center;
-  border: none;
-  background: none;
-  outline: none;
-
-  margin: 16px 0;
-  padding: 4px 0;
-
-  border-bottom: 1px solid #383838;
-}
-
 @keyframes vibrate {
   0% {
     transform: translateX(0);
